admin page question

+++ client/src/Pages/AdminPage/AdminPageContent.jsx
@@ -1,9 +1,14 @@
 import React, { useState } from "react";
 import PaginationComponent from "../../Components/PaginationComponent/PaginationComponent.jsx";
-import { updateProfile, addUser, deleteUser } from "../../Api/user-api";
+import { getUsers, updateProfile, addUser, deleteUser } from "../../Api/user-api";
 import { Pages } from "./AdminPage.jsx";
-import { getUsers } from "../../Api/user-api";
 import { useEffect } from "react";
+import {
+    getQuestions,
+    addQuestion,
+    updateQuestionAdmin,
+    deleteQuestion,
+} from "../../Api/question-api.js";
 
 // data props have at least one record to use following component
 function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPage }) {
@@ -17,6 +22,15 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
             getUsers(page, LimitRecordInPage).then((res) => {
                 setData(res.users);
             });
+        } else if (pageName === Pages.POST) {
+            getQuestions(page, LimitRecordInPage).then((res) => {
+                res.questions.forEach((question) => {
+                    delete question["usersLiked"];
+                    delete question["usersDisliked"];
+                    delete question["user"];
+                });
+                setData(res.questions);
+            });
         }
     }, [page]);
 
@@ -28,7 +42,21 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
                     (res) => res && alert("User updated successfully !!")
                 );
             } else {
+                delete currentFormData._id;
                 addUser(currentFormData).then((res) => res && alert("Add user success !!"));
+            }
+        } else if (pageName === Pages.POST) {
+            if (isFormUpdate) {
+                updateQuestionAdmin(currentFormData).then(
+                    (res) => res && alert("Question updated successfully !!")
+                );
+            } else {
+                delete currentFormData._id;
+                addQuestion(
+                    currentFormData.title,
+                    currentFormData.content,
+                    currentFormData.tags
+                ).then((res) => res && alert("Add question success !!"));
             }
         }
     }
@@ -38,7 +66,6 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
         if (!recordData?._id) {
             let recordDataEmpty = { ...data[0] };
             Object.keys(recordDataEmpty).forEach((key) => (recordDataEmpty[key] = ""));
-            delete recordDataEmpty._id;
             setCurrentFormData(recordDataEmpty);
             setIsFormUpdate(false);
         } else {
@@ -49,7 +76,13 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
     }
 
     function handleDeleteForm(e, _id) {
-        deleteUser(_id).then((res) => res && alert("User deleted !!"));
+        if (window.confirm("Are you sure you want to delete ?")) {
+            if (pageName === Pages.USER) {
+                deleteUser(_id).then((res) => res && alert("User deleted !!"));
+            } else if (pageName === Pages.POST) {
+                deleteQuestion(_id).then((res) => res && alert("Question deleted !!"));
+            }
+        }
     }
 
     function handleChangeForm(e, key) {
@@ -64,15 +97,15 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
             </div>
             <div className="main__list">
                 <div className="list__field">
-                    {Object.keys(data[0]).map((key, index) => (
-                        <span key={index}>{key}</span>
-                    ))}
+                    {Object.keys(data[0]).map(
+                        (key, index) => key !== "_id" && <span key={index}>{key}</span>
+                    )}
                 </div>
 
                 {data.map((obj, index) => (
                     <div key={index} className="list__record">
                         {Object.values(obj).map((value, index) => {
-                            return <span key={index}>{value}</span>;
+                            return index !== 0 && <span key={index}>{value}</span>;
                         })}
                         <button
                             onClick={(e) => {
@@ -100,16 +133,18 @@ function AdminPageContent({ data, setData, totalPages, pageName, LimitRecordInPa
                     <form onSubmit={handleSubmit}>
                         {Object.keys(currentFormData).map((key, index) => {
                             return (
-                                <React.Fragment key={index}>
-                                    <div className="form__title">{key}</div>
-                                    <input
-                                        type="text"
-                                        value={currentFormData[key]}
-                                        onChange={(e) => {
-                                            handleChangeForm(e, key);
-                                        }}
-                                    />
-                                </React.Fragment>
+                                key !== "_id" && (
+                                    <React.Fragment key={index}>
+                                        <div className="form__title">{key}</div>
+                                        <input
+                                            type="text"
+                                            value={currentFormData[key]}
+                                            onChange={(e) => {
+                                                handleChangeForm(e, key);
+                                            }}
+                                        />
+                                    </React.Fragment>
+                                )
                             );
                         })}
                         <input type="submit" value="Submit" />
